BACK, RANDOM,CURR c++ ok

+++ version_c/src/constants.h
@@ -3,12 +3,12 @@
 
 typedef struct _Config Config;
 typedef struct _Clustering Clustering;
-typedef struct _Clustering 
+typedef struct _Clustering
 {
-    double *p_c; // points coor
-    int *c_a;     // cluster assignements
-    int *n_p_p_c; // num_pts_per_clust
-    double *c_c; // cluster_centroids
+    float* p_c; // points coor
+    int* c_a;     // cluster assignements
+    int* n_p_p_c; // num_pts_per_clust
+    float* c_c; // cluster_centroids
 }Clustering;
 typedef struct _Config
 {
